feat(edc-api): restructure data-access libs for server feature
