--- libs/api/server-edc/data-access/project.json
@@ -1,8 +1,0 @@
-{
-  "name": "server-edc-data-access",
-  "$schema": "../../../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/api/server-edc/data-access/src",
-  "projectType": "library",
-  "targets": {},
-  "tags": ["type:feature", "scope:api"]
-}
