--- libs/api/server-emby/data-access/project.json
@@ -1,8 +1,0 @@
-{
-  "name": "server-emby-data-access",
-  "$schema": "../../../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/api/server-emby/data-access/src",
-  "projectType": "library",
-  "targets": {},
-  "tags": ["type:feature", "scope:api"]
-}
