+++ libs/api/server/data-access/emby/jest.config.ts
@@ -1,11 +1,11 @@
 /* eslint-disable */
 export default {
-  displayName: 'server-edc-data-access',
-  preset: '../../../../jest.preset.js',
+  displayName: 'server-data-access-emby',
+  preset: '../../../../../jest.preset.js',
   testEnvironment: 'node',
   transform: {
     '^.+\\.[tj]s$': ['ts-jest', { tsconfig: '<rootDir>/tsconfig.spec.json' }],
   },
   moduleFileExtensions: ['ts', 'js', 'html'],
-  coverageDirectory: '../../../../coverage/libs/api/server-edc/data-access',
+  coverageDirectory: '../../../../../coverage/libs/api/server/data-access/emby',
 };
