+++ libs/api/server/data-access/emby/project.json
@@ -1,0 +1,8 @@
+{
+  "name": "server-data-access-emby",
+  "$schema": "../../../../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "libs/api/server/data-access/emby/src",
+  "projectType": "library",
+  "targets": {},
+  "tags": ["type:feature", "scope:api"]
+}
