+++ libs/api/server/data-access/emby/src/index.ts
