+++ libs/api/server/feature/src/lib/server-feature.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { ServerFeatureController } from './server-feature.controller';
 import { ServerFeatureService } from './server-feature.service';
-import { ServerEdcDataAccessModule } from '@edc/api/server-edc/data-access';
-import { ServerEmbyDataAccessModule } from '@edc/api/server-emby/data-access';
+import { ServerEdcDataAccessModule } from '@edc/api/server/data-access/edc';
+import { ServerEmbyDataAccessModule } from '@edc/api/server/data-access/emby';
 
 @Module({
   imports: [ServerEdcDataAccessModule, ServerEmbyDataAccessModule],
