+++ libs/api/server/feature/src/lib/server-feature.service.ts
@@ -1,5 +1,5 @@
-import { ServerEdcDataAccessService } from '@edc/api/server-edc/data-access';
-import { ServerEmbyDataAccessService } from '@edc/api/server-emby/data-access';
+import { ServerEdcDataAccessService } from '@edc/api/server/data-access/edc';
+import { ServerEmbyDataAccessService } from '@edc/api/server/data-access/emby';
 import { Injectable } from '@nestjs/common';
 
 @Injectable()
