+++ tsconfig.base.json
@@ -19,8 +19,8 @@
       "@edc/api/mediaitem-edc/data-access": ["libs/api/mediaitem-edc/data-access/src/index.ts"],
       "@edc/api/mediaitem-emby/data-access": ["libs/api/mediaitem-emby/data-access/src/index.ts"],
       "@edc/api/mediaitem/feature": ["libs/api/mediaitem/feature/src/index.ts"],
-      "@edc/api/server-edc/data-access": ["libs/api/server-edc/data-access/src/index.ts"],
-      "@edc/api/server-emby/data-access": ["libs/api/server-emby/data-access/src/index.ts"],
+      "@edc/api/server/data-access/edc": ["libs/api/server/data-access/edc/src/index.ts"],
+      "@edc/api/server/data-access/emby": ["libs/api/server/data-access/emby/src/index.ts"],
       "@edc/api/server/feature": ["libs/api/server/feature/src/index.ts"],
       "@edc/api/user/data-access/edc": ["libs/api/user/data-access/edc/src/index.ts"],
       "@edc/api/user/data-access/emby": ["libs/api/user/data-access/emby/src/index.ts"],
